profevardillabackend/requirements.txt
langchain
langchain-huggingface
langchain-community
langchain-chroma
chromadb
pypdf
sentence-transformers
jsonschema
django
djangorestframework
django-cors-headers
pymupdf
tiktoken
rank_bm25
whitenoise
python-dotenv
groq
together
pillow
uvicorn
gunicorn
psutil
firebase-admin
django_ratelimit
torch==2.6.0